ml-service/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
numpy
pandas
pydantic
httpx
python-dotenv
pymongo
